Frag condition updates implemented

+++ phase4/src/updator.py
@@ -36,58 +36,65 @@ class Updator:
                     if j!=i and config.Horizontal_Fragments['Table_Name'][j] == config.parsedQuery.updateRelation:
                         altFrag = config.Horizontal_Fragments['Fragment_Name'][j]
                         altSite = config.Allocation['Site'][config.Allocation['Fragment_Name'].index(altFrag)]
-
-                        altCondition = config.Horizontal_Fragments['Attribute'][j]+config.Horizontal_Fragments['Operator'][j]+config.Horizontal_Fragments['Val'][j]
-                        print(altCondition)
+                        altCondition = ""
+                        if config.Horizontal_Fragments['Val'][j].isnumeric():
+                            altCondition = config.Horizontal_Fragments['Attribute'][j]+config.Horizontal_Fragments['Operator'][j]+config.Horizontal_Fragments['Val'][j]
+                        else:
+                            altCondition = config.Horizontal_Fragments['Attribute'][j]+config.Horizontal_Fragments['Operator'][j]+"'"+config.Horizontal_Fragments['Val'][j]+"'"
+                        config.debugPrint(altCondition)
 
                         nuTabName = altFrag+curFrag 
                         sqlQuery = "create table "+config.catalogName+"."+nuTabName+" select * from "+config.catalogName+"."+curFrag+" where "+altCondition+";"
-                        print(sqlQuery)
-                        # cur = config.globalConnections[curSite].cursor()
-                        # cur.execute(sqlQuery)
-                        # config.globalConnections[curSite].commit()
+                        # config.debugPrint("creation",sqlQuery)
+                        config.debugPrint(sqlQuery)
+                        cur = config.globalConnections[curSite].cursor()
+                        cur.execute(sqlQuery)
+                        config.globalConnections[curSite].commit()
 
                         sqlQuery = "DELETE FROM "+config.catalogName+"."+curFrag+" WHERE "+altCondition+";"
-                        print(sqlQuery)
-                        # cur = config.globalConnections[curSite].cursor()
-                        # cur.execute(sqlQuery)
-                        # config.globalConnections[curSite].commit()
+                        config.debugPrint(sqlQuery)
+                        cur = config.globalConnections[curSite].cursor()
+                        cur.execute(sqlQuery)
+                        config.globalConnections[curSite].commit()
 
-                        dumpTable(nuTabName,curSite)
-                        copyFromServer(curSite)
-                        copyToServer(altSite)
-                        importTable(altSite)
+                        if altSite!=curSite:
+                            dumpTable(nuTabName,curSite)
+                            copyFromServer(curSite)
+                            copyToServer(altSite)
+                            importTable(altSite)
 
                         finalTabName = nuTabName+"final"
                         sqlQuery = "Create Table "+config.catalogName+"."+finalTabName+" select * from "+config.catalogName+"."+altFrag+" UNION select * from "+config.catalogName+"."+nuTabName+";"
-                        print(sqlQuery)
-                        # cur = config.globalConnections[altSite].cursor()
-                        # cur.execute(sqlQuery)
-                        # config.globalConnections[altSite].commit()
+                        config.debugPrint(sqlQuery)
+                        cur = config.globalConnections[altSite].cursor()
+                        cur.execute(sqlQuery)
+                        config.globalConnections[altSite].commit()
 
-                        sqlQuery = "drop table "+config.catalogName+"."altFrag+";"
-                        print(sqlQuery)
-                        # cur = config.globalConnections[altSite].cursor()
-                        # cur.execute(sqlQuery)
-                        # config.globalConnections[altSite].commit()
+                        sqlQuery = "drop table "+config.catalogName+"."+altFrag+";"
+                        config.debugPrint(sqlQuery)
+                        cur = config.globalConnections[altSite].cursor()
+                        cur.execute(sqlQuery)
+                        config.globalConnections[altSite].commit()
 
                         sqlQuery = "alter table "+config.catalogName+"."+finalTabName+" RENAME "+config.catalogName+"."+altFrag+";"
-                        print(sqlQuery)
-                        # cur = config.globalConnections[altSite].cursor()
-                        # cur.execute(sqlQuery)
-                        # config.globalConnections[altSite].commit()
+                        config.debugPrint(sqlQuery)
+                        cur = config.globalConnections[altSite].cursor()
+                        cur.execute(sqlQuery)
+                        config.globalConnections[altSite].commit()
 
                         # remove nuTabName at curSite, altSite
                         sqlQuery = "drop table "+config.catalogName+"."+nuTabName+";"
-                        print(sqlQuery)
-                        # cur = config.globalConnections[altSite].cursor()
-                        # cur.execute(sqlQuery)
-                        # config.globalConnections[altSite].commit()
-                        sqlQuery = "drop table "+config.catalogName+"."+nuTabName+";"
-                        print(sqlQuery)
-                        # cur = config.globalConnections[curSite].cursor()
-                        # cur.execute(sqlQuery)
-                        # config.globalConnections[curSite].commit()
+                        config.debugPrint(sqlQuery)
+                        cur = config.globalConnections[altSite].cursor()
+                        cur.execute(sqlQuery)
+                        config.globalConnections[altSite].commit()
+                        if altSite != curSite:
+                            sqlQuery = "drop table "+config.catalogName+"."+nuTabName+";"
+                            config.debugPrint(sqlQuery)
+                            # config.debugPrint("final error querry",sqlQuery)
+                            cur = config.globalConnections[curSite].cursor()
+                            cur.execute(sqlQuery)
+                            config.globalConnections[curSite].commit()
 
         return
